Fix warnings. Closes #3656.
Patch from Mateusz Loskot.


[SVN r58011]

+++ src/engine/hcache.c
@@ -166,7 +166,7 @@ void write_netstring( FILE * f, char const * s )
 {
     if ( !s )
         s = "";
-    fprintf( f, "%lu\t%s\n", strlen( s ), s );
+    fprintf( f, "%lu\t%s\n", (long unsigned)strlen( s ), s );
 }
